App is working. * JobFragment writes to database * Need to read from database onStart
& ShiftsFragment reads & write data to/from firebase.

+++ app/src/main/java/myshift/com/UserPrefFb.java
@@ -2,9 +2,10 @@ package myshift.com;
 
 public class UserPrefFb {
 
-    public int hourlyRate;
+    public float hourlyRate;
     public int  breaks;
     public float bonusHours;
     public String jobName;
     public String jobAddress;
+
 }
